requirements.txt
streamlit
pandas
joblib
mysql-connector-python
textblob
scikit-learn
